diplomeBackend/diplomeBackend/src/main/java/com/example/diplomeBackend/services/UserService.java
package com.example.diplomeBackend.services;

import com.example.diplomeBackend.dto.UserDTO;
import com.example.diplomeBackend.exceptions.ResourceNotFoundException;
import com.example.diplomeBackend.mappers.UserMapper;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.stereotype.Service;
import com.example.diplomeBackend.models.User;
import com.example.diplomeBackend.repositories.UserRepository;

import java.util.List;
import java.util.Optional;
import java.util.stream.Collectors;

@Service
public class UserService {

    @Autowired
    private UserRepository userRepository;

    @Autowired
    private UserMapper userMapper;
    private final BCryptPasswordEncoder passwordEncoder = new BCryptPasswordEncoder();

    public List<User> findAll() {
        return userRepository.findAll();
    }

    public Optional<User> findById(Long id) {
        return userRepository.findById(id);
    }

    public User save(User user) {
        return userRepository.save(user);
    }

    public void deleteById(Long id) {
        userRepository.deleteById(id);
    }

    public User getUserById(Long id) {
        return userRepository.findById(id)
                .orElseThrow(() -> new RuntimeException("User not found with id: " + id));
    }


    public void saveUser(User user) {
        user.setPassword(hashPassword(user.getPassword()));
        userRepository.save(user);
}

    public String hashPassword(String password) {
        return passwordEncoder.encode(password);
    }

    public boolean checkPassword(String rawPassword, String encodedPassword) {
        return passwordEncoder.matches(rawPassword, encodedPassword);
    }
    public Optional<User> findByEmail(String email) {
        return userRepository.findByEmail(email);
    }

    public UserDTO findUserDTOById(Long userId) {
        User user = userRepository.findUserById(userId);
        return userMapper.toDTO(user);
    }

    public User findUserById(Long userId) {
        return userRepository.findUserById(userId);
    }

    public User updateUser(Long userId, User updatedUser) {
        return userRepository.findById(userId).map(user -> {
            user.setFirstName(updatedUser.getFirstName());
            user.setLastName(updatedUser.getLastName());
            user.setDateOfBirth(updatedUser.getDateOfBirth());
            user.setEmail(updatedUser.getEmail());
            user.setProfilePicture(updatedUser.getProfilePicture());
            user.setPassword(passwordEncoder.encode(updatedUser.getPassword()));
            return userRepository.save(user);
        }).orElseThrow(() -> new ResourceNotFoundException("User not found with id " + userId));
    }

    public List<UserDTO> searchUser(String query) {
        return userRepository.searchUser(query).stream()
                .map(userMapper::toDTO)
                .collect(Collectors.toList());
    }





}
